Add files via upload
Imagine standing in front of the entrance door to your office and seeing your face flash up on a large screen. If the screen identifies you, it opens the doors; if it doesn’t the doors don’t open. Your face is your digital ID. It allows you to enter your office and authenticate your identity in a hoard of other places. Sounds uncanny, but is undeniably remarkable! This not-so-old technology is on its way to becoming ubiquitous.


Enter the world of face detection and recognition systems…recognizing the face from all angles, with even more accuracy than a person can